Add run script user guide
Placing the run script user guide into the run script folder. It is a working document with changes to be made by those wo run the workflow, if they choose to use the wrapper scripts. It contains mostly commands for copying and pasting, rather than descriptive summaries. So it works in conjunction with the GDA2020 update process document which contains more description. The JADJ QA/QC and the NADJ sections still need refining.